+++ src/core/mmMoment.h
@@ -12,7 +12,7 @@ namespace MoMa {
         Moment( void );
         Moment( double time, double frameRate );
         Moment( unsigned int idx, double frameRate );
-        Moment( double time , unsigned int idx );
+        Moment( double time , unsigned long int idx );
         
         inline double time( void ) const { return( _time ); }
         inline unsigned int index( void ) const { return( _index ); }
@@ -21,18 +21,18 @@ namespace MoMa {
         void setFrameRate( double frameRate );
         
         void setTime( double time, double frameRate );
-        void setIndex( int idx, double frameRate );
+        void setIndex( long int idx, double frameRate );
         
         void setTime( double time );
-        void setIndex( unsigned int idx );
-        void setIndex( int idx );
+        void setIndex( unsigned long int idx );
+        void setIndex( long int idx );
 		Moment & operator++();
 		Moment & operator--();
         
       protected:
         
         double _frameRate;
-        unsigned int _index;
+        unsigned long int _index;
         double _time;
     };
 }
